updating userbased login

+++ src/posts/views.py
@@ -1,7 +1,7 @@
 from django.contrib import messages
 from django.core.paginator import Paginator,EmptyPage,PageNotAnInteger
 from django.shortcuts import render
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse,Http404,HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render, redirect
 import urllib.parse as parse
 
@@ -12,6 +12,9 @@ from .models import Post
 # Create your views here.
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -75,22 +78,26 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
-	instance = get_object_or_404(Post, slug=slug)
-	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
-	if form.is_valid():
-		instance = form.save(commit=False)
-		instance.save()
-		messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
-		return HttpResponseRedirect(instance.get_absolute_url())
-
-	context = {
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
+    form = PostForm(request.POST or None, request.FILES or None, instance=instance)
+    if form.is_valid():
+        instance = form.save(commit=False)
+        instance.save()
+        messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
+        return HttpResponseRedirect(instance.get_absolute_url())
+    context = {
 		"title": instance.title,
 		"instance": instance,
 		"form":form,
 	}
-	return render(request, "post_form.html", context)
+    return render(request, "post_form.html", context)
 
 def post_delete(request,slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
     instance = get_object_or_404(Post,slug=slug)
     instance.delete()
     messages.success(request,"Successfully deleted")
